fermium

Module audio

Audio playback.

Structs

SDL_AudioCVT

A structure to hold a set of audio conversion filters and buffers.

SDL_AudioDeviceID

Identifies an audio device.

SDL_AudioFormat

Audio format flags.

SDL_AudioSpec

The calculated values in this structure are calculated by SDL_OpenAudio().

SDL_AudioStatus

The status of an audio device.

SDL_AudioStream

SDL_AudioStream is a new audio conversion interface.

Constants

AUDIO_F32

Alias for AUDIO_F32LSB

AUDIO_F32LSB

32-bit floating point samples

AUDIO_F32MSB

As AUDIO_F32LSB, but big-endian byte order

AUDIO_F32SYS

Native-endian f32 samples

AUDIO_S8

Signed 8-bit samples

AUDIO_S16

Alias for AUDIO_S16LSB

AUDIO_S32

Alias for AUDIO_S32LSB;

AUDIO_S16LSB

Signed 16-bit samples

AUDIO_S16MSB

As AUDIO_U16MSB, but big-endian byte order

AUDIO_S16SYS

Native-endian i16 samples

AUDIO_S32LSB

32-bit integer samples

AUDIO_S32MSB

As AUDIO_S32MSB, but big-endian byte order

AUDIO_S32SYS

Native-endian i32 samples

AUDIO_U8

Unsigned 8-bit samples

AUDIO_U16

Alias for AUDIO_U16LSB

AUDIO_U16LSB

Unsigned 16-bit samples

AUDIO_U16MSB

As AUDIO_S16LSB, but big-endian byte order

AUDIO_U16SYS

Native-endian u16 samples

SDL_AUDIOCVT_MAX_FILTERS

You can only use 9 slots in the SDL_AudioCVT.filters field. The 10th is reserved as a null terminator.

SDL_AUDIO_PAUSED

SDL_AUDIO_PLAYING

SDL_AUDIO_STOPPED

SDL_MIX_MAXVOLUME

Maximum volume value, for use with SDL_MixAudio.

Functions

SDL_AUDIO_BITSIZE

Pits per sample. eg: i16 = 16, f32 = 32.

SDL_AUDIO_ISBIGENDIAN

If the samples are big-endian.

SDL_AUDIO_ISFLOAT

If the sample type is a floating type.

SDL_AUDIO_ISINT

If the samples are an int type.

SDL_AUDIO_ISLITTLEENDIAN

If the samples are little-endian.

SDL_AUDIO_ISSIGNED

If the samples are signed values.

SDL_AUDIO_ISUNSIGNED

If the samples are unsigned values.

SDL_AudioStreamAvailable⚠

Get the number of converted/re-sampled bytes available.

SDL_AudioStreamClear⚠

Clear any pending data in the stream without converting it

SDL_AudioStreamFlush⚠

Flush the stream.

SDL_AudioStreamGet⚠

Get converted/re-sampled data from the stream

SDL_AudioStreamPut⚠

Add data to be converted/re-sampled to the stream

SDL_BuildAudioCVT⚠

Initializes an SDL_AudioCVT.

SDL_ClearQueuedAudio⚠

Drop any queued audio data.

SDL_CloseAudio⚠

This function shuts down audio processing and closes the default audio device.

SDL_CloseAudioDevice⚠

This function shuts down audio processing and closes the given audio device.

SDL_ConvertAudio⚠

Performs an in-place audio conversion.

SDL_DequeueAudio⚠

Dequeue more audio on non-callback devices.

SDL_FreeAudioStream⚠

Free an audio stream

SDL_FreeWAV⚠

This function frees data previously allocated with SDL_LoadWAV_RW

SDL_GetAudioDeviceName⚠

Get the human-readable name of a specific audio device.

SDL_GetAudioDeviceSpec⚠

Get the preferred audio format of a specific audio device.

SDL_GetAudioDeviceStatus⚠

Gets the current audio status (stopped / playing / paused).

SDL_GetAudioDriver⚠

The name of an audio driver. See SDL_GetAudioDriver

SDL_GetAudioStatus⚠

Gets the current audio status (stopped / playing / paused).

SDL_GetCurrentAudioDriver⚠

The name of the current audio driver (or null if no driver is initialized).

SDL_GetNumAudioDevices⚠

Get the number of available devices exposed by the current driver.

SDL_GetNumAudioDrivers⚠

The number of audio drivers. See SDL_GetAudioDriver

SDL_GetQueuedAudioSize⚠

Get the number of bytes of still-queued audio.

SDL_LoadWAV⚠

Loads a WAV file.

SDL_LoadWAV_RW⚠

Load the audio data of a WAVE file into memory

SDL_LockAudio⚠

Locks the default audio device.

SDL_LockAudioDevice⚠

Locks the given audio device.

SDL_MixAudio⚠

In-place mixes two audio sources.

SDL_MixAudioFormat⚠

Mix according to a format.

SDL_NewAudioStream⚠

Create a new audio stream

SDL_OpenAudio⚠

This function opens the audio device with the desired parameters, and returns 0 if successful, placing the actual hardware parameters in the structure pointed to by obtained. If obtained is NULL, the audio data passed to the callback function will be guaranteed to be in the requested format, and will be automatically converted to the hardware audio format if necessary. This function returns -1 if it failed to open the audio device, or couldn’t set up the audio thread.

SDL_OpenAudioDevice⚠

Open a specific audio device.

SDL_PauseAudio⚠

Pause and unpause the audio callback processing.

SDL_PauseAudioDevice⚠

Pause and unpause the audio callback processing.

SDL_QueueAudio⚠

Queue more audio on non-callback devices.

SDL_UnlockAudio⚠

Unlocks the default audio device.

SDL_UnlockAudioDevice⚠

Unlocks the given audio device.

Type Aliases

SDL_AudioCallback

This function is called when the audio device needs more data.

SDL_AudioFilter

Filter function used by SDL_AudioCVT