fermium::audio

Function SDL_MixAudio

pub unsafe extern "C" fn SDL_MixAudio(
    dst: *mut u8,
    src: *const u8,
    len: u32,
    volume: c_int,
)

In-place mixes two audio sources.

This takes two audio buffers of the playing audio format and mixes them, performing addition, volume adjustment, and overflow clipping. The volume ranges from 0 - 128, and should be set to SDL_MIX_MAXVOLUME for full audio volume. Note this does not change hardware volume.

This is provided for convenience, you can mix your own audio data.