fermium::audio

Struct SDL_AudioCVT

#[repr(C)]
pub struct SDL_AudioCVT {
    pub needed: c_int,
    pub src_format: SDL_AudioFormat,
    pub dst_format: SDL_AudioFormat,
    pub rate_incr: f64,
    pub buf: *mut u8,
    pub len: c_int,
    pub len_cvt: c_int,
    pub len_mult: c_int,
    pub len_ratio: f64,
    pub filters: [SDL_AudioFilter; 10],
    pub filter_index: c_int,
}

A structure to hold a set of audio conversion filters and buffers.

Note that various parts of the conversion pipeline can take advantage of SIMD operations (like SSE2, for example). SDL_AudioCVT doesn’t require you to pass it aligned data, but can possibly run much faster if you set both its (buf) field to a pointer that is aligned to 16 bytes, and its (len) field to something that’s a multiple of 16, if possible.

Fields
needed: c_int

Set to 1 if conversion possible

src_format: SDL_AudioFormat

Source audio format

dst_format: SDL_AudioFormat

Target audio format

rate_incr: f64

Rate conversion increment

buf: *mut u8

Buffer to hold entire audio data

len: c_int

Length of original audio buffer

len_cvt: c_int

Length of converted audio buffer

len_mult: c_int

buffer must be len*len_mult big

len_ratio: f64

Given len, final size is len*len_ratio

filters: [SDL_AudioFilter; 10]

NULL-terminated list of filter functions

filter_index: c_int

Current audio conversion function