fermium

Module events

Module for event handling.

Structs

SDL_AudioDeviceEvent

Audio device event structure (event.adevice.*)

SDL_CommonEvent

Fields shared by every event.

SDL_ControllerAxisEvent

Game controller axis motion event structure (event.caxis.*)

SDL_ControllerButtonEvent

Game controller button event structure (event.cbutton.*)

SDL_ControllerDeviceEvent

Controller device event structure (event.cdevice.*)

SDL_ControllerSensorEvent

Game controller sensor event structure (event.csensor.*)

SDL_ControllerTouchpadEvent

Game controller touchpad event structure (event.ctouchpad.*)

SDL_DisplayEvent

Display state change event data (event.display.*)

SDL_DollarGestureEvent

Dollar Gesture Event (event.dgesture.*)

SDL_DropEvent

An event used to request a file open by the system (event.drop.*)

SDL_EventType

The types of events that can be delivered.

SDL_JoyAxisEvent

Joystick axis motion event structure (event.jaxis.*)

SDL_JoyBallEvent

Joystick trackball motion event structure (event.jball.*)

SDL_JoyButtonEvent

Joystick button event structure (event.jbutton.*)

SDL_JoyDeviceEvent

Joystick device event structure (event.jdevice.*)

SDL_JoyHatEvent

Joystick hat position change event structure (event.jhat.*)

SDL_KeyboardEvent

Keyboard button event structure (event.key.*)

SDL_MouseButtonEvent

Mouse button event structure (event.button.*)

SDL_MouseMotionEvent

Mouse motion event structure (event.motion.*)

SDL_MouseWheelEvent

Mouse wheel event structure (event.wheel.*)

SDL_MultiGestureEvent

Multiple Finger Gesture Event (event.mgesture.*)

SDL_OSEvent

OS Specific event.

SDL_QuitEvent

The “quit requested” event

SDL_SensorEvent

Sensor event structure (event.sensor.*)

SDL_SysWMEvent

A video driver dependent system event (event.syswm.*)

SDL_SysWMmsg

Message info from the system’s window manager.

SDL_SysWMmsg_windows

This is the data your Window Procedure would get.

SDL_TextEditingEvent

Keyboard text editing event structure (event.edit.*)

SDL_TextInputEvent

Keyboard text input event structure (event.text.*)

SDL_TouchFingerEvent

Touch finger event structure (event.tfinger.*)

SDL_UserEvent

A user-defined event type (event.user.*)

SDL_WindowEvent

Window state change event data (event.window.*)

SDL_eventaction

Used with SDL_PeepEvents

Constants

SDL_ADDEVENT

SDL_APP_DIDENTERBACKGROUND

The application did enter the background and may not get CPU for some time.

SDL_APP_DIDENTERFOREGROUND

The application is now interactive.

SDL_APP_LOWMEMORY

The application is low on memory, free memory if possible.

SDL_APP_TERMINATING

The application is being terminated by the OS.

SDL_APP_WILLENTERBACKGROUND

The application is about to enter the background.

SDL_APP_WILLENTERFOREGROUND

The application is about to enter the foreground.

SDL_AUDIODEVICEADDED

A new audio device is available.

SDL_AUDIODEVICEREMOVED

An audio device has been removed.

SDL_CLIPBOARDUPDATE

The clipboard changed.

SDL_CONTROLLERAXISMOTION

Game controller axis motion.

SDL_CONTROLLERBUTTONDOWN

Game controller button pressed.

SDL_CONTROLLERBUTTONUP

Game controller button released.

SDL_CONTROLLERDEVICEADDED

A new Game controller has been inserted into the system.

SDL_CONTROLLERDEVICEREMAPPED

The controller mapping was updated.

SDL_CONTROLLERDEVICEREMOVED

An opened Game controller has been removed.

SDL_CONTROLLERSENSORUPDATE

Game controller sensor was updated.

SDL_CONTROLLERTOUCHPADDOWN

Game controller touchpad was touched.

SDL_CONTROLLERTOUCHPADMOTION

Game controller touchpad finger was moved.

SDL_CONTROLLERTOUCHPADUP

Game controller touchpad finger was lifted.

SDL_DISABLE

Alias for SDL_IGNORE

SDL_DISPLAYEVENT

Display state change.

SDL_DOLLARGESTURE

SDL_DOLLARRECORD

SDL_DROPBEGIN

A new set of drops is beginning (NULL filename).

SDL_DROPCOMPLETE

Current set of drops is now complete (NULL filename).

SDL_DROPFILE

The system requests a file open.

SDL_DROPTEXT

Text/plain drag-and-drop event.

SDL_ENABLE

Used with SDL_EventState

SDL_FINGERDOWN

SDL_FINGERMOTION

SDL_FINGERUP

SDL_FIRSTEVENT

Unused

SDL_GETEVENT

SDL_IGNORE

Used with SDL_EventState

SDL_JOYAXISMOTION

Joystick axis motion.

SDL_JOYBALLMOTION

Joystick trackball motion.

SDL_JOYBUTTONDOWN

Joystick button pressed.

SDL_JOYBUTTONUP

Joystick button released.

SDL_JOYDEVICEADDED

A new joystick has been inserted into the system.

SDL_JOYDEVICEREMOVED

An opened joystick has been removed.

SDL_JOYHATMOTION

Joystick hat position change.

SDL_KEYDOWN

Key pressed.

SDL_KEYMAPCHANGED

Keymap changed due to a system event such as an input language or keyboard layout change.

SDL_KEYUP

Key released.

SDL_LASTEVENT

This last event is only for bounding internal arrays.

SDL_LOCALECHANGED

The user’s locale preferences have changed.

SDL_MOUSEBUTTONDOWN

Mouse button pressed.

SDL_MOUSEBUTTONUP

Mouse button released.

SDL_MOUSEMOTION

Mouse moved.

SDL_MOUSEWHEEL

Mouse wheel motion.

SDL_MULTIGESTURE

SDL_PEEKEVENT

SDL_PRESSED

Button is pressed.

SDL_QUERY

Used with SDL_EventState

SDL_QUIT

User-requested quit

SDL_RELEASED

Button is released.

SDL_RENDER_DEVICE_RESET

The device has been reset and all textures need to be recreated.

SDL_RENDER_TARGETS_RESET

The render targets have been reset and their contents need to be updated.

SDL_SENSORUPDATE

A sensor was updated.

SDL_SYSWMEVENT

System specific event.

SDL_TEXTEDITING

Keyboard text editing (composition).

SDL_TEXTEDITINGEVENT_TEXT_SIZE

Size of the SDL_TextEditingEvent array.

SDL_TEXTINPUT

Keyboard text input.

SDL_TEXTINPUTEVENT_TEXT_SIZE

Size of the SDL_TextInputEvent array.

SDL_USEREVENT

Events SDL_USEREVENT through SDL_LASTEVENT are for your use, and should be allocated with SDL_RegisterEvents.

SDL_WINDOWEVENT

Window state change.

Functions

SDL_AddEventWatch⚠

Add a function which is called when an event is added to the queue.

SDL_DelEventWatch⚠

Remove an event watch function added with SDL_AddEventWatch

SDL_EventState⚠

This function allows you to set the state of processing certain events.

SDL_FilterEvents⚠

Run the filter function on the current event queue, removing any events for which the filter returns 0.

SDL_FlushEvent⚠

This function clears events of a particular type from the event queue.

SDL_FlushEvents⚠

This function clears a range of event types from the event queue.

SDL_GetEventFilter⚠

Return the current event filter - can be used to “chain” filters.

SDL_GetEventState⚠

As SDL_EventState, but provides the “query” part for you.

SDL_HasEvent⚠

Checks to see if certain event types are in the event queue.

SDL_HasEvents⚠

Checks to see if a range of event types are in the event queue.

SDL_PeepEvents⚠

Checks the event queue for messages and optionally returns them.

SDL_PollEvent⚠

Polls for currently pending events.

SDL_PumpEvents⚠

Pumps the event loop, gathering events from the input devices.

SDL_PushEvent⚠

Add an event to the event queue.

SDL_RegisterEvents⚠

This function allocates a set of user-defined events, and returns the beginning event number for that set of events.

SDL_SetEventFilter⚠

Sets up a filter to process all events before they change internal state and are posted to the internal event queue.

SDL_WaitEvent⚠

Waits indefinitely for the next available event.

SDL_WaitEventTimeout⚠

Waits until the specified timeout (in milliseconds) for the next available event.

Type Aliases

SDL_EventFilter

Event filter operation.

Unions

SDL_Event

General event structure

SDL_SysWMmsg_union