fermium::gamecontroller

Function SDL_GameControllerRumble

pub unsafe extern "C" fn SDL_GameControllerRumble(
    gamecontroller: *mut SDL_GameController,
    low_frequency_rumble: u16,
    high_frequency_rumble: u16,
    duration_ms: u32,
) -> c_int

Trigger a rumble effect.

Each call to this function cancels any previous rumble effect, and calling it with 0 intensity stops any rumbling.

gamecontroller The controller to vibrate

low_frequency_rumble The intensity of the low frequency (left) rumble motor, from 0 to 0xFFFF

high_frequency_rumble The intensity of the high frequency (right) rumble motor, from 0 to 0xFFFF

duration_ms The duration of the rumble effect, in milliseconds

Return: 0, or -1 if rumble isn’t supported on this controller