fermium::joystick

Function SDL_JoystickEventState

pub unsafe extern "C" fn SDL_JoystickEventState(state: c_int) -> c_int

Enable/disable joystick event polling.

If joystick events are disabled, you must call SDL_JoystickUpdate yourself and check the state of the joystick when you want joystick information.

The state can be one of SDL_QUERY, SDL_ENABLE, or SDL_IGNORE.