fermium::joystick

Function SDL_JoystickRumble

pub unsafe extern "C" fn SDL_JoystickRumble(
    joystick: *mut SDL_Joystick,
    low_frequency_rumble: u16,
    high_frequency_rumble: u16,
    duration_ms: u32,
) -> c_int

Start a rumble effect.

Each call to this function cancels any previous rumble effect, and calling it with 0 intensity stops any rumbling.

joystick The joystick to vibrate.

low_frequency_rumble The intensity of the low frequency (left) rumble motor, from 0 to 0xFFFF.

high_frequency_rumble The intensity of the high frequency (right) rumble motor, from 0 to 0xFFFF.

duration_ms The duration of the rumble effect, in milliseconds.

Returns: 0, or -1 if rumble isn’t supported on this joystick