fermium

Module keyboard

Interactions with the keyboard.

Structs

SDL_Keysym

SDL uses this for key events.

Functions

SDL_GetKeyFromName⚠

Get a key code from a human-readable name.

SDL_GetKeyFromScancode⚠

Get the key code corresponding to the given scancode according to the current keyboard layout.

SDL_GetKeyName⚠

Get a human-readable name for a key.

SDL_GetKeyboardFocus⚠

Get the window which currently has keyboard focus.

SDL_GetKeyboardState⚠

Get a snapshot of the current state of the keyboard.

SDL_GetModState⚠

Get the current key modifier state for the keyboard.

SDL_GetScancodeFromKey⚠

Get the scancode corresponding to the given key code according to the current keyboard layout.

SDL_GetScancodeFromName⚠

Get a scancode from a human-readable name.

SDL_GetScancodeName⚠

Get a human-readable name for a scancode.

SDL_HasScreenKeyboardSupport⚠

Returns whether the platform has some screen keyboard support.

SDL_IsScreenKeyboardShown⚠

Returns whether the screen keyboard is shown for given window.

SDL_IsTextInputActive⚠

Return whether or not Unicode text input events are enabled.

SDL_SetModState⚠

Set the current key modifier state for the keyboard.

SDL_SetTextInputRect⚠

Set the rectangle used to type Unicode text inputs.

SDL_StartTextInput⚠

Start accepting Unicode text input events.

SDL_StopTextInput⚠

Stop receiving any text input events.