fermium

Module messagebox

Allows the creation of message boxes, for short messages to the user.

Structs

SDL_MessageBoxButtonData

Individual button data.

SDL_MessageBoxButtonFlags

SDL_MessageBoxButtonData flags.

SDL_MessageBoxColor

RGB value used in a message box color scheme

SDL_MessageBoxColorScheme

A set of colors to use for message box dialogs

SDL_MessageBoxColorType

See the SDL_MESSAGEBOX_COLOR_* constants.

SDL_MessageBoxData

MessageBox structure containing title, text, window, etc.

SDL_MessageBoxFlags

Flags for SDL_ShowMessageBox and SDL_ShowSimpleMessageBox.

Constants

SDL_MESSAGEBOX_BUTTONS_LEFT_TO_RIGHT

buttons placed left to right

SDL_MESSAGEBOX_BUTTONS_RIGHT_TO_LEFT

buttons placed right to left

SDL_MESSAGEBOX_BUTTON_ESCAPEKEY_DEFAULT

Marks the default button when escape is hit.

SDL_MESSAGEBOX_BUTTON_RETURNKEY_DEFAULT

Marks the default button when return is hit.

SDL_MESSAGEBOX_COLOR_BACKGROUND

SDL_MESSAGEBOX_COLOR_BUTTON_BACKGROUND

SDL_MESSAGEBOX_COLOR_BUTTON_BORDER

SDL_MESSAGEBOX_COLOR_BUTTON_SELECTED

SDL_MESSAGEBOX_COLOR_MAX

SDL_MESSAGEBOX_COLOR_TEXT

SDL_MESSAGEBOX_ERROR

error dialog

SDL_MESSAGEBOX_INFORMATION

informational dialog

SDL_MESSAGEBOX_WARNING

warning dialog

Functions

SDL_ShowMessageBox⚠

Create a modal message box.

SDL_ShowSimpleMessageBox⚠

Create a simple modal message box.