fermium

Module mouse

Mouse control and interaction.

Structs

SDL_Cursor

An opaque cursor value.

SDL_MouseWheelDirection

Scroll direction types for the Scroll event

SDL_SystemCursor

Used with SDL_CreateSystemCursor

Constants

SDL_BUTTON_LEFT

Individual value for the left button.

SDL_BUTTON_LMASK

Mask for the left button.

SDL_BUTTON_MIDDLE

Individual value for the middle button (mouse wheel).

SDL_BUTTON_MMASK

Mask for the middle button (mouse wheel).

SDL_BUTTON_RIGHT

Individual value for the right button.

SDL_BUTTON_RMASK

Mask for the right button.

SDL_BUTTON_X1

Individual value for the extra button 1.

SDL_BUTTON_X2

Individual value for the extra button 2.

SDL_BUTTON_X1MASK

Mask for the extra button 1.

SDL_BUTTON_X2MASK

Mask for the extra button 2.

SDL_MOUSEWHEEL_FLIPPED

The scroll direction is flipped / natural

SDL_MOUSEWHEEL_NORMAL

The scroll direction is normal

SDL_NUM_SYSTEM_CURSORS

Used for array size purposes and such.

SDL_SYSTEM_CURSOR_ARROW

Arrow

SDL_SYSTEM_CURSOR_CROSSHAIR

Crosshair

SDL_SYSTEM_CURSOR_HAND

Hand

SDL_SYSTEM_CURSOR_IBEAM

I-beam

SDL_SYSTEM_CURSOR_NO

Slashed circle or crossbones

SDL_SYSTEM_CURSOR_SIZEALL

Four pointed arrow pointing north, south, east, and west

SDL_SYSTEM_CURSOR_SIZENESW

Double arrow pointing northeast and southwest

SDL_SYSTEM_CURSOR_SIZENS

Double arrow pointing north and south

SDL_SYSTEM_CURSOR_SIZENWSE

Double arrow pointing northwest and southeast

SDL_SYSTEM_CURSOR_SIZEWE

Double arrow pointing west and east

SDL_SYSTEM_CURSOR_WAIT

Wait

SDL_SYSTEM_CURSOR_WAITARROW

Small wait cursor (or Wait if not available)

Functions

SDL_BUTTON

Used as a mask when testing buttons in buttonstate.

SDL_CaptureMouse⚠

Capture the mouse, to track input outside an SDL window.

SDL_CreateColorCursor⚠

Create a color cursor.

SDL_CreateCursor⚠

Create a cursor, using the specified bitmap data and mask (in MSB format).

SDL_CreateSystemCursor⚠

Create a system cursor.

SDL_FreeCursor⚠

Frees a cursor created with SDL_CreateCursor or similar functions.

SDL_GetCursor⚠

Return the active cursor.

SDL_GetDefaultCursor⚠

Return the default cursor.

SDL_GetGlobalMouseState⚠

Get the current state of the mouse, in relation to the desktop.

SDL_GetMouseFocus⚠

Get the window which currently has mouse focus.

SDL_GetMouseState⚠

Retrieve the current state of the mouse.

SDL_GetRelativeMouseMode⚠

Query whether relative mouse mode is enabled.

SDL_GetRelativeMouseState⚠

Retrieve the relative state of the mouse.

SDL_SetCursor⚠

Set the active cursor.

SDL_SetRelativeMouseMode⚠

Set relative mouse mode.

SDL_ShowCursor⚠

Toggle whether or not the cursor is shown.

SDL_WarpMouseGlobal⚠

Moves the mouse to the given position in global screen space.

SDL_WarpMouseInWindow⚠

Moves the mouse to the given position within the window.