fermium

Module pixels

Module for pixel editing.

Structs

SDL_ArrayOrder

Array component order, low byte -> high byte.

SDL_BitmapOrder

Bitmap pixel order, high bit -> low bit.

SDL_Color

An RGBA color value (8-bits per channel).

SDL_PackedLayout

Packed component order, high bit -> low bit.

SDL_PackedOrder

Packed component order, high bit -> low bit.

SDL_Palette

Info about a palette of colors.

SDL_PixelFormat

An SDL Pixel Format.

SDL_PixelFormatEnum

An enumerated pixel format value.

SDL_PixelType

Pixel type.

Constants

SDL_ALPHA_OPAQUE

More alpha => more opaque.

SDL_ALPHA_TRANSPARENT

Transparent pixels are alpha = 0.

SDL_ARRAYORDER_ABGR

SDL_ARRAYORDER_ARGB

SDL_ARRAYORDER_BGR

SDL_ARRAYORDER_BGRA

SDL_ARRAYORDER_NONE

SDL_ARRAYORDER_RGB

SDL_ARRAYORDER_RGBA

SDL_BITMAPORDER_1234

Bitmap ordering from low to high.

SDL_BITMAPORDER_4321

Bitmap ordering from high to low.

SDL_BITMAPORDER_NONE

No bitmap ordering.

SDL_PACKEDLAYOUT_332

SDL_PACKEDLAYOUT_565

SDL_PACKEDLAYOUT_1555

SDL_PACKEDLAYOUT_4444

SDL_PACKEDLAYOUT_5551

SDL_PACKEDLAYOUT_8888

SDL_PACKEDLAYOUT_1010102

SDL_PACKEDLAYOUT_2101010

SDL_PACKEDLAYOUT_NONE

SDL_PACKEDORDER_ABGR

SDL_PACKEDORDER_ARGB

SDL_PACKEDORDER_BGRA

SDL_PACKEDORDER_BGRX

SDL_PACKEDORDER_NONE

SDL_PACKEDORDER_RGBA

SDL_PACKEDORDER_RGBX

SDL_PACKEDORDER_XBGR

SDL_PACKEDORDER_XRGB

SDL_PIXELFORMAT_ABGR32

SDL_PIXELFORMAT_ABGR1555

SDL_PIXELFORMAT_ABGR4444

SDL_PIXELFORMAT_ABGR8888

SDL_PIXELFORMAT_ARGB32

SDL_PIXELFORMAT_ARGB1555

SDL_PIXELFORMAT_ARGB4444

SDL_PIXELFORMAT_ARGB8888

SDL_PIXELFORMAT_ARGB2101010

SDL_PIXELFORMAT_BGR24

SDL_PIXELFORMAT_BGR444

SDL_PIXELFORMAT_BGR555

SDL_PIXELFORMAT_BGR565

SDL_PIXELFORMAT_BGR888

SDL_PIXELFORMAT_BGRA32

SDL_PIXELFORMAT_BGRA4444

SDL_PIXELFORMAT_BGRA5551

SDL_PIXELFORMAT_BGRA8888

SDL_PIXELFORMAT_BGRX8888

SDL_PIXELFORMAT_EXTERNAL_OES

Android video texture format

SDL_PIXELFORMAT_INDEX8

SDL_PIXELFORMAT_INDEX1LSB

SDL_PIXELFORMAT_INDEX1MSB

SDL_PIXELFORMAT_INDEX4LSB

SDL_PIXELFORMAT_INDEX4MSB

SDL_PIXELFORMAT_IYUV

Planar mode: Y + U + V (3 planes)

SDL_PIXELFORMAT_NV12

Planar mode: Y + U/V interleaved (2 planes)

SDL_PIXELFORMAT_NV21

Planar mode: Y + V/U interleaved (2 planes)

SDL_PIXELFORMAT_RGB24

SDL_PIXELFORMAT_RGB332

SDL_PIXELFORMAT_RGB444

SDL_PIXELFORMAT_RGB555

SDL_PIXELFORMAT_RGB565

SDL_PIXELFORMAT_RGB888

SDL_PIXELFORMAT_RGBA32

SDL_PIXELFORMAT_RGBA4444

SDL_PIXELFORMAT_RGBA5551

SDL_PIXELFORMAT_RGBA8888

SDL_PIXELFORMAT_RGBX8888

SDL_PIXELFORMAT_UNKNOWN

SDL_PIXELFORMAT_UYVY

Packed mode: U0+Y0+V0+Y1 (1 plane)

SDL_PIXELFORMAT_XBGR1555

SDL_PIXELFORMAT_XBGR4444

SDL_PIXELFORMAT_XBGR8888

SDL_PIXELFORMAT_XRGB1555

SDL_PIXELFORMAT_XRGB4444

SDL_PIXELFORMAT_XRGB8888

SDL_PIXELFORMAT_YUY2

Packed mode: Y0+U0+Y1+V0 (1 plane)

SDL_PIXELFORMAT_YV12

Planar mode: Y + V + U (3 planes)

SDL_PIXELFORMAT_YVYU

Packed mode: Y0+V0+Y1+U0 (1 plane)

SDL_PIXELTYPE_ARRAYF16

SDL_PIXELTYPE_ARRAYF32

SDL_PIXELTYPE_ARRAYU8

SDL_PIXELTYPE_ARRAYU16

SDL_PIXELTYPE_ARRAYU32

SDL_PIXELTYPE_INDEX1

SDL_PIXELTYPE_INDEX4

SDL_PIXELTYPE_INDEX8

SDL_PIXELTYPE_PACKED8

SDL_PIXELTYPE_PACKED16

SDL_PIXELTYPE_PACKED32

SDL_PIXELTYPE_UNKNOWN

An unknown pixel type.

Functions

SDL_AllocFormat⚠

Create an SDL_PixelFormat structure from a pixel format enum.

SDL_AllocPalette⚠

Create a palette structure with the specified number of color entries.

SDL_BITSPERPIXEL

Gets the bits per pixel.

SDL_BYTESPERPIXEL

Gets the bytes per pixel.

SDL_CalculateGammaRamp⚠

Calculate a 256 entry gamma ramp for a gamma value.

SDL_DEFINE_PIXELFORMAT

Converts the pixel type parameters into a pixel format enumeration value.

SDL_DEFINE_PIXELFOURCC

Converts a FourCC into a pixel format enumeration value.

SDL_FreeFormat⚠

Free an SDL_PixelFormat structure.

SDL_FreePalette⚠

Free a palette created with SDL_AllocPalette.

SDL_GetPixelFormatName⚠

Get the human readable name of a pixel format

SDL_GetRGB⚠

Get the RGB components from a pixel of the specified format.

SDL_GetRGBA⚠

Get the RGBA components from a pixel of the specified format.

SDL_ISPIXELFORMAT_ALPHA

Does the pixel format have an alpha channel?

SDL_ISPIXELFORMAT_ARRAY

Is this pixel format an array format?

SDL_ISPIXELFORMAT_FOURCC

Is the pixel format a FourCC format?

SDL_ISPIXELFORMAT_INDEXED

Is this pixel format an indexed format?

SDL_ISPIXELFORMAT_PACKED

Is this pixel format a packed format?

SDL_MapRGB⚠

Maps an RGB triple to an opaque pixel value for a given pixel format.

SDL_MapRGBA⚠

Maps an RGBA quadruple to a pixel value for a given pixel format.

SDL_MasksToPixelFormatEnum⚠

Convert a bpp and RGBA masks to an enumerated pixel format.

SDL_PIXELFLAG

Gets the pixel flag bits.

SDL_PIXELLAYOUT

Gets the pixel layout bits.

SDL_PIXELORDER

Gets the pixel order bits.

SDL_PIXELTYPE

Gets the pixel type bits.

SDL_PixelFormatEnumToMasks⚠

Convert one of the enumerated pixel formats to a bpp and RGBA masks.

SDL_SetPaletteColors⚠

Set a range of colors in a palette.

SDL_SetPixelFormatPalette⚠

Set the palette for a pixel format structure.