fermium::renderer

Function SDL_CreateTextureFromSurface

pub unsafe extern "C" fn SDL_CreateTextureFromSurface(
    renderer: *mut SDL_Renderer,
    surface: *mut SDL_Surface,
) -> *mut SDL_Texture

Create a texture from an existing surface.

renderer The renderer.

surface The surface containing pixel data used to fill the texture.

Returns: The created texture is returned, or NULL on error.

The surface is not modified or freed by this function.

See Also: SDL_QueryTexture, SDL_DestroyTexture