fermium::renderer

Function SDL_CreateWindowAndRenderer

pub unsafe extern "C" fn SDL_CreateWindowAndRenderer(
    width: c_int,
    height: c_int,
    window_flags: u32,
    window: *mut *mut SDL_Window,
    renderer: *mut *mut SDL_Renderer,
) -> c_int

Create a window and default renderer

width The width of the window

height The height of the window

window_flags The flags used to create the window

window A pointer filled with the window, or NULL on error

renderer A pointer filled with the renderer, or NULL on error

Returns: 0 on success, or -1 on error