fermium::renderer

Function SDL_RenderCopyEx

pub unsafe extern "C" fn SDL_RenderCopyEx(
    renderer: *mut SDL_Renderer,
    texture: *mut SDL_Texture,
    srcrect: *const SDL_Rect,
    dstrect: *const SDL_Rect,
    angle: f64,
    center: *const SDL_Point,
    flip: SDL_RendererFlip,
) -> c_int

Copy a portion of the source texture to the current rendering target, rotating it by angle around the given center

renderer The renderer which should copy parts of a texture.

texture The source texture.

srcrect A pointer to the source rectangle, or NULL for the entire texture.

dstrect A pointer to the destination rectangle, or NULL for the entire rendering target.

angle An angle in degrees that indicates the rotation that will be applied to dstrect, rotating it in a clockwise direction

center A pointer to a point indicating the point around which dstrect will be rotated (if NULL, rotation will be done around dstrect.w/2, dstrect.h/2).

flip An SDL_RendererFlip value stating which flipping actions should be performed on the texture

Returns: 0 on success, or -1 on error