fermium

Module surface

Module for SDL’s in-memory bitmap type.

Structs

SDL_BlitMap

A newtype’d c_void.

SDL_Surface

Represents a bitmap image in CPU memory.

SDL_YUV_CONVERSION_MODE

The formula used for converting between YUV and RGB.

Constants

SDL_DONTFREE

Surface is referenced internally

SDL_PREALLOC

Surface uses preallocated memory

SDL_RLEACCEL

Surface is RLE encoded

SDL_SIMD_ALIGNED

Surface uses aligned memory

SDL_YUV_CONVERSION_AUTOMATIC

BT.601 for SD content, BT.709 for HD content

SDL_YUV_CONVERSION_BT601

BT.601 (the default)

SDL_YUV_CONVERSION_BT709

BT.709

SDL_YUV_CONVERSION_JPEG

Full range JPEG

Functions

SDL_BlitScaled⚠

This is the public scaled blit function, SDL_BlitScaled(), and it performs rectangle validation and clipping before passing it to SDL_LowerBlitScaled()

SDL_BlitSurface⚠

Performs a fast blit from the source surface to the destination surface.

SDL_ConvertPixels⚠

Copy a block of pixels of one format to another format

SDL_ConvertSurface⚠

Creates a new surface of the specified format, and then copies and maps the given surface to it so the blit of the converted surface will be as fast as possible. If this function fails, it returns NULL.

SDL_ConvertSurfaceFormat⚠

See SDL_ConvertSurface

SDL_CreateRGBSurface⚠

Allocate a new RGB surface.

SDL_CreateRGBSurfaceFrom⚠

Makes a surface from a pre-allocated buffer.

SDL_CreateRGBSurfaceWithFormat⚠

Similar to SDL_CreateRGBSurface, but using a pixel format enum value.

SDL_CreateRGBSurfaceWithFormatFrom⚠

Makes a surface from a pre-allocated buffer in a specified format.

SDL_DuplicateSurface⚠

Creates a new surface identical to the existing surface

SDL_FillRect⚠

Performs a fast fill of the given rectangle with color.

SDL_FillRects⚠

See SDL_FillRect

SDL_FreeSurface⚠

Frees the surface.

SDL_GetClipRect⚠

Gets the clipping rectangle for the destination surface in a blit.

SDL_GetColorKey⚠

Gets the color key (transparent pixel) in a blit-able surface.

SDL_GetSurfaceBlendMode⚠

Get the blend mode used for blit operations.

SDL_GetSurfaceColorMod⚠

Get the additional color value used in blit operations.

SDL_GetYUVConversionMode⚠

Get the YUV conversion mode

SDL_GetYUVConversionModeForResolution⚠

Get the YUV conversion mode.

SDL_HasColorKey⚠

Returns whether the surface has a color key

SDL_HasSurfaceRLE⚠

Returns whether the surface is RLE enabled

SDL_LoadBMP⚠

Given a file path to a BMP file, load and parse it into an SDL_Surface as one step.

SDL_LoadBMP_RW⚠

Load a surface from a seekable SDL data stream (memory or file).

SDL_LockSurface⚠

Sets up a surface for directly accessing the pixels.

SDL_MUSTLOCK⚠

Checks if an SDL_Surface must be locked before the pixels are accessed directly.

SDL_SaveBMP⚠

Save an SDL_Surface to the file path specified.

SDL_SaveBMP_RW⚠

Save a surface to a seekable SDL data stream (memory or file).

SDL_SetClipRect⚠

Sets the clipping rectangle for the destination surface in a blit.

SDL_SetColorKey⚠

Sets the color key (transparent pixel) in a blit-able surface.

SDL_SetSurfaceAlphaMod⚠

Get the additional alpha value used in blit operations.

SDL_SetSurfaceBlendMode⚠

Set the blend mode used for blit operations.

SDL_SetSurfaceColorMod⚠

Set an additional color value used in blit operations.

SDL_SetSurfacePalette⚠

Set the palette used by a surface.

SDL_SetSurfaceRLE⚠

Sets the RLE acceleration hint for a surface.

SDL_SetYUVConversionMode⚠

Set the YUV conversion mode

SDL_SoftStretchLinear⚠

Perform bilinear scaling between two surfaces of the same format, 32BPP.

SDL_UnlockSurface⚠

See SDL_LockSurface