fermium

Module timer

Timer related functionality.

Structs

SDL_TimerID

Definition of the timer ID type.

Functions

SDL_AddTimer⚠

Add a new timer to the pool of timers already running.

SDL_Delay⚠

Wait a specified number of milliseconds before returning.

SDL_GetPerformanceCounter⚠

Get the current value of the high resolution counter.

SDL_GetPerformanceFrequency⚠

Get the count per second of the high resolution counter.

SDL_GetTicks⚠

Get the number of milliseconds since the SDL library initialization.

SDL_RemoveTimer⚠

Remove a timer by ID.

Type Aliases

SDL_TimerCallback

Function prototype for the timer callback function.