fermium

Module touch

Module for touch interface interactions.

Structs

SDL_Finger

Info for a finger object.

SDL_FingerID

Used to identify a finger object.

SDL_TouchDeviceType

The types of a touch device.

SDL_TouchID

Used with touch events.

Constants

SDL_MOUSE_TOUCHID

Used as the SDL_TouchID for touch events simulated with mouse input.

SDL_TOUCH_DEVICE_DIRECT

Touch screen with window-relative coordinates

SDL_TOUCH_DEVICE_INDIRECT_ABSOLUTE

Trackpad with absolute device coordinates

SDL_TOUCH_DEVICE_INDIRECT_RELATIVE

Trackpad with screen cursor-relative coordinates

SDL_TOUCH_DEVICE_INVALID

The “invalid” touch type. Used for error values and such.

SDL_TOUCH_MOUSEID

Used as the device ID for mouse events simulated with touch input.

Functions

SDL_GetNumTouchDevices⚠

Get the number of registered touch devices.

SDL_GetNumTouchFingers⚠

Get the number of active fingers for a given touch device.

SDL_GetTouchDevice⚠

Get the touch ID with the given index, or 0 if the index is invalid.

SDL_GetTouchDeviceType⚠

Get the type of the given touch device.

SDL_GetTouchFinger⚠

Get the finger object of the given touch, with the given index.