fermium

Module video

Module for interacting with the video subsystem.

Structs

SDL_DisplayEventID

Event subtype for display events.

SDL_DisplayMode

The structure that defines a display mode

SDL_DisplayOrientation

Orientations a display can have.

SDL_FlashOperation

Window flash operation

SDL_GLContext

A handle to an OpenGL context.

SDL_GLContextResetNotification

Affects GL’s behavior when the context is reset.

SDL_GLattr

OpenGL configuration attributes

SDL_GLcontextFlag

SDL Context Flags.

SDL_GLcontextReleaseFlag

Affects GL’s behavior when you release the context.

SDL_GLprofile

The GL Profile: Core, Compatibility, or ES.

SDL_HitTestResult

The results of a hit test.

SDL_Window

The type used to identify a window (newtype’d c_void).

SDL_WindowEventID

Event subtype for window events.

SDL_WindowFlags

The flags on a window

Constants

SDL_DISPLAYEVENT_CONNECTED

Display has been added to the system

SDL_DISPLAYEVENT_DISCONNECTED

Display has been removed from the system

SDL_DISPLAYEVENT_NONE

Never used

SDL_DISPLAYEVENT_ORIENTATION

Display orientation has changed to data1

SDL_FLASH_BRIEFLY

Flash the window briefly to get attention

SDL_FLASH_CANCEL

Cancel any window flash state

SDL_FLASH_UNTIL_FOCUSED

Flash the window until it gets focus

SDL_GL_ACCELERATED_VISUAL

SDL_GL_ACCUM_ALPHA_SIZE

SDL_GL_ACCUM_BLUE_SIZE

SDL_GL_ACCUM_GREEN_SIZE

SDL_GL_ACCUM_RED_SIZE

SDL_GL_ALPHA_SIZE

SDL_GL_BLUE_SIZE

SDL_GL_BUFFER_SIZE

SDL_GL_CONTEXT_DEBUG_FLAG

SDL_GL_CONTEXT_EGL

SDL_GL_CONTEXT_FLAGS

SDL_GL_CONTEXT_FORWARD_COMPATIBLE_FLAG

SDL_GL_CONTEXT_MAJOR_VERSION

SDL_GL_CONTEXT_MINOR_VERSION

SDL_GL_CONTEXT_NO_ERROR

SDL_GL_CONTEXT_PROFILE_COMPATIBILITY

SDL_GL_CONTEXT_PROFILE_CORE

SDL_GL_CONTEXT_PROFILE_ES

SDL_GL_CONTEXT_PROFILE_MASK

SDL_GL_CONTEXT_RELEASE_BEHAVIOR

SDL_GL_CONTEXT_RELEASE_BEHAVIOR_FLUSH

SDL_GL_CONTEXT_RELEASE_BEHAVIOR_NONE

SDL_GL_CONTEXT_RESET_ISOLATION_FLAG

SDL_GL_CONTEXT_RESET_LOSE_CONTEXT

SDL_GL_CONTEXT_RESET_NOTIFICATION

SDL_GL_CONTEXT_RESET_NO_NOTIFICATION

SDL_GL_CONTEXT_ROBUST_ACCESS_FLAG

SDL_GL_DEPTH_SIZE

SDL_GL_DOUBLEBUFFER

SDL_GL_FRAMEBUFFER_SRGB_CAPABLE

SDL_GL_GREEN_SIZE

SDL_GL_MULTISAMPLEBUFFERS

SDL_GL_MULTISAMPLESAMPLES

SDL_GL_RED_SIZE

SDL_GL_RETAINED_BACKING

SDL_GL_SHARE_WITH_CURRENT_CONTEXT

SDL_GL_STENCIL_SIZE

SDL_GL_STEREO

SDL_HITTEST_DRAGGABLE

SDL_HITTEST_NORMAL

SDL_HITTEST_RESIZE_BOTTOM

SDL_HITTEST_RESIZE_BOTTOMLEFT

SDL_HITTEST_RESIZE_BOTTOMRIGHT

SDL_HITTEST_RESIZE_LEFT

SDL_HITTEST_RESIZE_RIGHT

SDL_HITTEST_RESIZE_TOP

SDL_HITTEST_RESIZE_TOPLEFT

SDL_HITTEST_RESIZE_TOPRIGHT

SDL_ORIENTATION_LANDSCAPE

SDL_ORIENTATION_LANDSCAPE_FLIPPED

SDL_ORIENTATION_PORTRAIT

SDL_ORIENTATION_PORTRAIT_FLIPPED

SDL_ORIENTATION_UNKNOWN

SDL_WINDOWEVENT_CLOSE

SDL_WINDOWEVENT_ENTER

SDL_WINDOWEVENT_EXPOSED

SDL_WINDOWEVENT_FOCUS_GAINED

SDL_WINDOWEVENT_FOCUS_LOST

SDL_WINDOWEVENT_HIDDEN

SDL_WINDOWEVENT_HIT_TEST

SDL_WINDOWEVENT_LEAVE

SDL_WINDOWEVENT_MAXIMIZED

SDL_WINDOWEVENT_MINIMIZED

SDL_WINDOWEVENT_MOVED

SDL_WINDOWEVENT_NONE

SDL_WINDOWEVENT_RESIZED

SDL_WINDOWEVENT_RESTORED

SDL_WINDOWEVENT_SHOWN

SDL_WINDOWEVENT_SIZE_CHANGED

SDL_WINDOWEVENT_TAKE_FOCUS

SDL_WINDOWPOS_CENTERED

Specifies that you want the window centered.

SDL_WINDOWPOS_UNDEFINED

Specifies that you don’t care about the window position.

SDL_WINDOW_ALLOW_HIGHDPI

SDL_WINDOW_ALWAYS_ON_TOP

SDL_WINDOW_BORDERLESS

SDL_WINDOW_FOREIGN

SDL_WINDOW_FULLSCREEN

SDL_WINDOW_FULLSCREEN_DESKTOP

SDL_WINDOW_HIDDEN

SDL_WINDOW_INPUT_FOCUS

SDL_WINDOW_INPUT_GRABBED

SDL_WINDOW_MAXIMIZED

SDL_WINDOW_METAL

SDL_WINDOW_MINIMIZED

SDL_WINDOW_MOUSE_CAPTURE

SDL_WINDOW_MOUSE_FOCUS

SDL_WINDOW_OPENGL

SDL_WINDOW_POPUP_MENU

SDL_WINDOW_RESIZABLE

SDL_WINDOW_SHOWN

SDL_WINDOW_SKIP_TASKBAR

SDL_WINDOW_TOOLTIP

SDL_WINDOW_UTILITY

SDL_WINDOW_VULKAN

Functions

SDL_CreateWindow⚠

Create a window with the specified position, dimensions, and flags.

SDL_CreateWindowFrom⚠

Creates a window from driver-dependent window creation data, typically a native window pointer.

SDL_DestroyWindow⚠

Destroys a window.

SDL_DisableScreenSaver⚠

Disables the screensaver.

SDL_EnableScreenSaver⚠

Enables the screensaver.

SDL_FlashWindow⚠

Request a window to demand attention from the user.

SDL_GL_CreateContext⚠

Creates a new context for use with the window, and also makes it current.

SDL_GL_DeleteContext⚠

Deletes a GL context.

SDL_GL_ExtensionSupported⚠

If the given extension is supported in the current context.

SDL_GL_GetAttribute⚠

Gets a GL attribute’s value (to the pointer given).

SDL_GL_GetCurrentContext⚠

Gets the current context.

SDL_GL_GetCurrentWindow⚠

Gets the window associated with the current context.

SDL_GL_GetDrawableSize⚠

Gets the drawable size (in pixels) of the client area.

SDL_GL_GetProcAddress⚠

Get a GL function pointer for a given function name.

SDL_GL_GetSwapInterval⚠

Gets the swap interval setting.

SDL_GL_LoadLibrary⚠

Load a GL library using the file path given.

SDL_GL_MakeCurrent⚠

Make a context current and associated with the given window.

SDL_GL_ResetAttributes⚠

Resets all GL attributes to their default values.

SDL_GL_SetAttribute⚠

Sets a GL attribute to the given value.

SDL_GL_SetSwapInterval⚠

Sets the swap interval of GL swaps.

SDL_GL_SwapWindow⚠

Swaps the front buffer and back buffer of a GL using window.

SDL_GL_UnloadLibrary⚠

Unloads the previously loaded GL library.

SDL_GetClosestDisplayMode⚠

Gets the closest display mode to the mode requested.

SDL_GetCurrentDisplayMode⚠

Gets info about the current display mode.

SDL_GetCurrentVideoDriver⚠

The name of the current video driver.

SDL_GetDesktopDisplayMode⚠

Gets info about the desktop’s display mode.

SDL_GetDisplayBounds⚠

Writes the bounds of a given display to the rect given.

SDL_GetDisplayDPI⚠

Get the diagonal, horizontal, and vertical DPI of a display.

SDL_GetDisplayMode⚠

Gets info about a display mode.

SDL_GetDisplayName⚠

Gets the name of a display.

SDL_GetDisplayOrientation⚠

Gets the orientation of a display.

SDL_GetDisplayUsableBounds⚠

As SDL_GetDisplayBounds, but system reserved area (eg: the taskbar) is removed for you.

SDL_GetGrabbedWindow⚠

Gets which window has the mouse grab (if any).

SDL_GetNumDisplayModes⚠

Gets the number of available display modes.

SDL_GetNumVideoDisplays⚠

The number of displays available.

SDL_GetNumVideoDrivers⚠

The number of available video drivers.

SDL_GetVideoDriver⚠

Lookup the name of a video driver.

SDL_GetWindowBordersSize⚠

Gets the size of the border decoration around the client area.

SDL_GetWindowBrightness⚠

Gets the brightness (gamma) of the display that owns the window.

SDL_GetWindowData⚠

Gets the named user data pointer of the window.

SDL_GetWindowDisplayIndex⚠

Gets the display index for the center of a given window.

SDL_GetWindowDisplayMode⚠

Gets the display mode used when the given window is visible and fullscreen.

SDL_GetWindowFlags⚠

Gets the flags of the window.

SDL_GetWindowFromID⚠

Get a window from a stored ID, or NULL if it doesn’t exist.

SDL_GetWindowGammaRamp⚠

Gets the gamma ramp of the window.

SDL_GetWindowGrab⚠

Gets if the window is grabbing the mouse.

SDL_GetWindowID⚠

Gets the window’s ID, or 0 on failure.

SDL_GetWindowKeyboardGrab⚠

Get a window’s keyboard grab mode.

SDL_GetWindowMaximumSize⚠

Gets the maximum client area size of the window.

SDL_GetWindowMinimumSize⚠

Gets the minimum client area size of the window.

SDL_GetWindowMouseGrab⚠

Get a window’s mouse grab mode.

SDL_GetWindowOpacity⚠

Gets a window’s opacity (to the pointer given).

SDL_GetWindowPixelFormat⚠

Gets the pixel format of the window.

SDL_GetWindowPosition⚠

Gets the window’s position.

SDL_GetWindowSize⚠

Gets the window’s client area size (in screen coordinates).

SDL_GetWindowSurface⚠

Gets the surface of the window.

SDL_GetWindowTitle⚠

Gets the window’s title, or "" if there is no title.

SDL_HideWindow⚠

Hides a window.

SDL_IsScreenSaverEnabled⚠

If the screensaver is currently enabled.

SDL_MaximizeWindow⚠

Maximize a window.

SDL_MinimizeWindow⚠

Minimize a window.

SDL_RaiseWindow⚠

Raises a window to the front and sets it for input focus.

SDL_RestoreWindow⚠

Restores a minimized/maximized window to its previous size and position.

SDL_SetWindowAlwaysOnTop⚠

Set the window to always be above the others.

SDL_SetWindowBordered⚠

Add/remove the border of the window.

SDL_SetWindowBrightness⚠

Set the brightness (gamma) for the display that owns the window.

SDL_SetWindowData⚠

Associates a named user data pointer to the window.

SDL_SetWindowDisplayMode⚠

Sets the display mode to use when the given window is visible and fullscreen.

SDL_SetWindowFullscreen⚠

Sets the window’s desired fullscreen state.

SDL_SetWindowGammaRamp⚠

Sets a gamma ramp for the window.

SDL_SetWindowGrab⚠

Sets if the window “grabs” the mouse pointer (locking it inside the window).

SDL_SetWindowHitTest⚠

Sets the hit test callback hit test user data for the window.

SDL_SetWindowIcon⚠

Sets the window’s icon.

SDL_SetWindowInputFocus⚠

Sets input focus to a given window.

SDL_SetWindowKeyboardGrab⚠

Set a window’s keyboard grab mode.

SDL_SetWindowMaximumSize⚠

Sets the maximum client area size of the window.

SDL_SetWindowMinimumSize⚠

Sets the minimum client area size of the window.

SDL_SetWindowModalFor⚠

Sets a window as a modal for another window (X11 only).

SDL_SetWindowMouseGrab⚠

Set a window’s mouse grab mode.

SDL_SetWindowOpacity⚠

Sets a window’s opacity.

SDL_SetWindowPosition⚠

Sets the window’s position.

SDL_SetWindowResizable⚠

Set if the window should allow resizing.

SDL_SetWindowSize⚠

Sets the window’s client area size (in screen coordinates).

SDL_SetWindowTitle⚠

Sets the title of a window.

SDL_ShowWindow⚠

Shows a window.

SDL_UpdateWindowSurface⚠

Updates the window surface data for the user to see.

SDL_UpdateWindowSurfaceRects⚠

As SDL_UpdateWindowSurfaceRects, but limited to the list of rects given.

SDL_VideoInit⚠

Initialize the video subsystem (and other subsystems it depends on)

SDL_VideoQuit⚠

Quit the video subsystem (and also other event subsystems)

Type Aliases

SDL_HitTest

SDL’s “hit test” function pointer type.